requirements.txt
fastapi
uvicorn
openai
python-dotenv
asyncio
pydantic